refactor: PaginatedResponse 타입을 shared/api 레이어로 이동 (FSD 준수)
features 레이어 간 직접 참조는 FSD 아키텍처 위반이므로
범용 HTTP 타입인 PaginatedResponse를 shared/api로 이동.

- shared/api/client.ts: PaginatedResponse<T> 단일 정의 추가
- features/resume/api/types.ts: 로컬 정의 제거, shared에서 re-export
- features/interview-session/api/types.ts: 동일
- features/interview-setup/api/interviewSetupApi.ts: @/features/resume → @/shared/api
- features/user-job-description/api/userJobDescriptionApi.ts: 동일 (이전 커밋에 포함)

의존 방향: pages → features → shared (올바른 FSD 단방향 의존)
