+++ src/features/interview-session/api/types.ts
@@ -83,13 +83,8 @@ export interface SubmitAnswerFollowupResponse {
 export type SubmitAnswerFullProcessResponse = InterviewTurn | { detail: string };
 export type SubmitAnswerResponse = SubmitAnswerFollowupResponse | SubmitAnswerFullProcessResponse;
 
-export interface PaginatedResponse<T> {
-  count: number;
-  totalPagesCount: number;
-  nextPage: number | null;
-  previousPage: number | null;
-  results: T[];
-}
+// ── Paginated Response (공통 타입 — shared 에서 re-export) ──
+export type { PaginatedResponse } from "@/shared/api/client";
 
 export interface InterviewSessionListItem {
   uuid: string;
